Add a file filtered_logger.py that contains a function filter_datum
Implemented a function filter_datum that accepts four Arguments. fields a list of strings representing all fields to obfuscate, redaction: a string representing by what the field will be obfuscated, message: a string representing the log line, separator: a string representing by which character is separating all fields in the log line (message). The function returns the updated verion of the message